Split theme (common css parsing code) and media query function codes
