+++ pkg/lib/theme/theme.go
@@ -2,11 +2,8 @@ package theme
 
 import (
 	"bytes"
-	"errors"
-	"io"
 	"strings"
 
-	"github.com/tdewolff/parse/v2"
 	"github.com/tdewolff/parse/v2/css"
 )
 
@@ -71,32 +68,6 @@ func (r *atrule) Stringify(buf *bytes.Buffer, indent indentation) {
 	}
 	buf.Write([]byte(indent))
 	buf.Write([]byte("}\n"))
-}
-
-// MigrateMediaQueryToClassBased migrates media query dark theme to class-based dark theme.
-func MigrateMediaQueryToClassBased(r io.Reader) (result []byte, err error) {
-	p := css.NewParser(parse.NewInput(r), false)
-
-	var elements []element
-	for {
-		var el element
-		el, err = parseElement(p)
-		if errors.Is(err, io.EOF) {
-			err = nil
-			break
-		}
-		if err != nil {
-			return
-		}
-		elements = append(elements, el)
-	}
-
-	elements = transform(elements)
-	var buf bytes.Buffer
-	stringify(&buf, elements)
-
-	result = buf.Bytes()
-	return
 }
 
 func parseAtrule(p *css.Parser, a *atrule) (err error) {
@@ -195,27 +166,6 @@ func collectTokensAsString(tokens []css.Token) string {
 		buf.Write(token.Data)
 	}
 	return buf.String()
-}
-
-func transform(elements []element) (out []element) {
-	for _, el := range elements {
-		switch v := el.(type) {
-		case *atrule:
-			if v.Identifier == "@media" && v.Value == "(prefers-color-scheme:dark)" {
-				// Remove this at rule
-				for _, ruleset := range v.Rulesets {
-					if ruleset.Selector == ":root" {
-						ruleset.Selector = ":root.dark"
-					}
-					out = append(out, ruleset)
-				}
-			}
-		default:
-			out = append(out, el)
-		}
-	}
-
-	return
 }
 
 func stringify(buf *bytes.Buffer, elements []element) {
